requirements.txt
djangorestframework-simplejwt